Refactor githubprovider package and remove unused code
